@zhuzilin 2021-11-25
In GPT2_4B model, the first 5 loss values are:
- before this MR: [0.763671875, 7.359375, 11.9921875, 7.18359375, 12.6484375]
- this MR: [0.763671875, 7.359375, 11.9921875, 7.18359375, 12.6484375]
- before #243: [0.763671875, nan, nan, nan, nan]

This MR makes sure the numerical correctness by ensuring the computation relates to the about-to-move chunk is finished.