@zhuzilin 2021-11-25
Note the code update here has an error. `ChunkState.COMPUTE` should be `TensorState.COMPUTE`. The benchmark was also updated.